requirements.txt
numpy
pandas
scikit-learn
python-telegram-bot
matplotlib
xgboost
schedule
requests